# lib/api-zod/src/generated/types/aiSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabilGen Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { AiSettingsProvider } from "./aiSettingsProvider";

export interface AiSettings {
  /** Which AI provider to use for generation */
  provider: AiSettingsProvider;
  /**
   * DeepSeek API key
   * @nullable
   */
  deepseekApiKey: string | null;
  /** DeepSeek API base URL */
  deepseekBaseUrl: string;
  /** DeepSeek model to use (e.g. deepseek-chat) */
  deepseekModel: string;
}